PUB-2359 Removed calls to channel-management (#447)
* Removed calls to channel-management

* Bumping chart version/ fixing aliases

* Fixed checkstyle

* Update readme diagram

* Update format of publication file endpoints

* Bumping chart version/ fixing aliases

---------
