+++ Jenkinsfile_CNP
@@ -20,7 +20,6 @@ def setupTestSecrets() {
       secret('app-pip-subscription-management-scope', 'SUBSCRIPTION_MANAGEMENT_AZ_API'),
       secret('app-pip-publication-services-scope', 'PUBLICATION_SERVICES_AZ_API'),
       secret('app-pip-account-management-scope', 'ACCOUNT_MANAGEMENT_AZ_API'),
-      secret('app-pip-channel-management-scope', 'CHANNEL_MANAGEMENT_AZ_API'),
       secret('app-tenant', 'TENANT_ID')
     ]) {
     env.NOTIFY_API_KEY = "${NOTIFY_API_KEY}"
@@ -29,7 +28,6 @@ def setupTestSecrets() {
     env.CLIENT_SECRET = "${CLIENT_SECRET}"
     env.DATA_MANAGEMENT_AZ_API = "${DATA_MANAGEMENT_AZ_API}"
     env.ACCOUNT_MANAGEMENT_AZ_API = "${ACCOUNT_MANAGEMENT_AZ_API}"
-    env.CHANNEL_MANAGEMENT_AZ_API = "${CHANNEL_MANAGEMENT_AZ_API}"
     env.PUBLICATION_SERVICES_AZ_API = "${PUBLICATION_SERVICES_AZ_API}"
     env.SUBSCRIPTION_MANAGEMENT_AZ_API = "${SUBSCRIPTION_MANAGEMENT_AZ_API}"
     env.TENANT_ID = "${TENANT_ID}"
