+++ src/main/java/uk/gov/hmcts/reform/pip/publication/services/service/DataManagementService.java
@@ -9,9 +9,13 @@ import org.springframework.web.reactive.function.client.WebClient;
 import org.springframework.web.reactive.function.client.WebClientResponseException;
 import uk.gov.hmcts.reform.pip.model.location.Location;
 import uk.gov.hmcts.reform.pip.model.publication.Artefact;
+import uk.gov.hmcts.reform.pip.model.publication.FileType;
 import uk.gov.hmcts.reform.pip.publication.services.errorhandling.exceptions.ServiceToServiceException;
 
 import java.util.UUID;
+
+import static org.springframework.http.HttpStatus.NOT_FOUND;
+import static org.springframework.http.HttpStatus.PAYLOAD_TOO_LARGE;
 
 @Slf4j
 @Service
@@ -20,7 +24,10 @@ public class DataManagementService {
 
     private static final String SERVICE = "Data Management";
     private static final String ADMIN_HEADER = "x-admin";
+    private static final String SYSTEM_HEADER = "x-system";
+    private static final String ADDITIONAL_PDF_HEADER = "x-additional-pdf";
     private static final String TRUE = "true";
+    private static final int MAX_FILE_SIZE = 2_000_000;
 
     @Value("${service-to-service.data-management}")
     private String url;
@@ -72,6 +79,36 @@ public class DataManagementService {
                 .accept(MediaType.APPLICATION_JSON)
                 .retrieve().bodyToMono(String.class).block();
         } catch (WebClientResponseException ex) {
+            throw new ServiceToServiceException(SERVICE, ex.getMessage());
+        }
+    }
+
+    public String getArtefactSummary(UUID artefactId) {
+        try {
+            return webClient.get().uri(String.format("%s/publication/%s/summary", url, artefactId))
+                .accept(MediaType.APPLICATION_JSON)
+                .retrieve().bodyToMono(String.class)
+                .block();
+        } catch (WebClientResponseException ex) {
+            throw new ServiceToServiceException(SERVICE, ex.getMessage());
+        }
+    }
+
+    public String getArtefactFile(UUID artefactId, FileType fileType, boolean additionalPdf) {
+        try {
+            return webClient.get()
+                .uri(String.format("%s/publication/%s/%s?maxFileSize=%s", url, artefactId, fileType, MAX_FILE_SIZE))
+                .header(SYSTEM_HEADER, TRUE)
+                .header(ADDITIONAL_PDF_HEADER, String.valueOf(additionalPdf))
+                .accept(MediaType.APPLICATION_JSON)
+                .retrieve()
+                .bodyToMono(String.class)
+                .block();
+        } catch (WebClientResponseException ex) {
+            if (NOT_FOUND.equals(ex.getStatusCode())
+                || PAYLOAD_TOO_LARGE.equals(ex.getStatusCode())) {
+                return "";
+            }
             throw new ServiceToServiceException(SERVICE, ex.getMessage());
         }
     }
