+++ src/main/resources/application.yaml
@@ -32,10 +32,6 @@ spring:
             authorization-grant-type: client_credentials
             scopes:
               - ${DATA_MANAGEMENT_AZ_API}/.default
-          channelManagementApi:
-            authorization-grant-type: client_credentials
-            scopes:
-              - ${CHANNEL_MANAGEMENT_AZ_API}/.default
           subscriptionManagementApi:
             authorization-grant-type: client_credentials
             scopes:
@@ -51,7 +47,6 @@ spring:
 
 service-to-service:
   data-management: ${DATA_MANAGEMENT_URL:https://pip-data-management.staging.platform.hmcts.net}
-  channel-management: ${CHANNEL_MANAGEMENT_URL:https://pip-channel-management.staging.platform.hmcts.net}
   subscription-management: ${SUBSCRIPTION_MANAGEMENT_URL:https://pip-subscription-management.staging.platform.hmcts.net}
   account-management: ${ACCOUNT_MANAGEMENT_URL:https://pip-account-management.staging.platform.hmcts.net}
