+++ src/test/java/uk/gov/hmcts/reform/pip/publication/services/service/SubscriptionNotificationServiceTest.java
@@ -78,9 +78,6 @@ class SubscriptionNotificationServiceTest extends RedisConfigurationTestBase {
     @MockBean
     private DataManagementService dataManagementService;
 
-    @MockBean
-    private ChannelManagementService channelManagementService;
-
     @Autowired
     private SubscriptionNotificationService notificationService;
 
@@ -113,10 +110,10 @@ class SubscriptionNotificationServiceTest extends RedisConfigurationTestBase {
         when(sendEmailResponse.getReference()).thenReturn(Optional.of(SUCCESS_REF_ID));
         when(emailService.sendEmail(any())).thenReturn(sendEmailResponse);
 
-        when(channelManagementService.getArtefactFile(ARTEFACT_ID, FileType.PDF, false)).thenReturn(FILE_CONTENT);
-        when(channelManagementService.getArtefactFile(ARTEFACT_ID, FileType.PDF, true)).thenReturn(FILE_CONTENT);
-        when(channelManagementService.getArtefactSummary(ARTEFACT_ID)).thenReturn(ARTEFACT_SUMMARY);
-        when(channelManagementService.getArtefactFile(ARTEFACT_ID, FileType.EXCEL, false)).thenReturn(FILE_CONTENT);
+        when(dataManagementService.getArtefactFile(ARTEFACT_ID, FileType.PDF, false)).thenReturn(FILE_CONTENT);
+        when(dataManagementService.getArtefactFile(ARTEFACT_ID, FileType.PDF, true)).thenReturn(FILE_CONTENT);
+        when(dataManagementService.getArtefactSummary(ARTEFACT_ID)).thenReturn(ARTEFACT_SUMMARY);
+        when(dataManagementService.getArtefactFile(ARTEFACT_ID, FileType.EXCEL, false)).thenReturn(FILE_CONTENT);
         when(dataManagementService.getArtefactFlatFile(ARTEFACT_ID)).thenReturn(ARTEFACT_FLAT_FILE);
     }
